requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=/root/work/abantether/requirements.txt /root/work/abantether/requirements.in
#
beautifulsoup4==4.12.2
    # via -r /root/work/abantether/requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
loguru==0.7.0
    # via -r /root/work/abantether/requirements.in
python-dotenv==1.0.0
    # via -r /root/work/abantether/requirements.in
requests==2.30.0
    # via -r /root/work/abantether/requirements.in
soupsieve==2.4.1
    # via beautifulsoup4
urllib3==2.0.2
    # via requests
